Move compare headers into CompareItem
